requirements.txt
pymupdf
faiss-cpu
numpy
flask
openai
python-
sentence-transformers